conditional rendering of assignmentview renaming for accuracy
